+++ home.html
@@ -352,6 +352,6 @@
       crossorigin="anonymous"
     ></script>
     <script src="../../src/js/script.mjs" type="module"></script>
-    <script src="./src/js/homepage/homepageScript.js" type="module"></script>
+    <script src="./src/js/homepageScript.js" type="module"></script>
   </body>
 </html>
