+++ profile.html
@@ -300,6 +300,6 @@
       crossorigin="anonymous"
     ></script>
     <script src="../../src/js/script.mjs" type="module"></script>
-    <script src="./src/js/profile/profileScript.mjs" type="module"></script>
+    <script src="./src/js/profileScript.mjs" type="module"></script>
   </body>
 </html>
